// src/graphql/balancer/codegen.tsx
export type BalancerLbpQuery = { pools: Array<{ id: string, swapsCount: string, holdersCount: string, tokens?: Array<{ id: string, symbol: string, balance: string, weight?: string | null }> | null }>, latestPrice?: { id: string, price: string, pricingAsset: string, block: string } | null };